agent/pyproject.toml
[project]
name = "text2sql-agent"
version = "0.1.0"
description = "Text 2 SQL Reasoning Engine using LangGraph"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-postgres>=0.0.16",  # For pgvector
    "psycopg2-binary>=2.9.0",     # DB Driver
    "langchain-mcp-adapters>=0.1.0",  # For connecting to Phase 1 Server
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["agent_core*"]
